Rename settings files to conform to camelCasing
