+++ src/settings/index.ts
@@ -18,9 +18,9 @@ import defaultMapIcon from "assets/icons/default-map-icon.png";
 import sparkleNavLogo from "assets/icons/sparkle-nav-logo.png";
 import sparkleverseLogo from "assets/images/sparkleverse-logo.png";
 
-export * from "./url-settings";
-export * from "./use-settings-defaults";
-export * from "./portal-settings";
+export * from "./urlSettings";
+export * from "./useSettingsDefaults";
+export * from "./portalSettings";
 
 export const SPARKLE_HOMEPAGE_URL = "https://sparklespaces.com/";
 export const SPARKLE_TERMS_AND_CONDITIONS_URL =
